+++ adapters/__init__.py
@@ -2,11 +2,7 @@
 from .filesystem_adapter import LocalFileSystemAdapter
 from .embedding import MockEmbeddingAdapter, OpenAIEmbeddingAdapter
 from .desktop_adapter import MockDesktopAdapter, PyAutoGUIAdapter
+from .agent_adapter import RuleBasedAgentAdapter
 
-__all__ = [
-    "LocalFileSystemAdapter",
-    "MockEmbeddingAdapter",
-    "OpenAIEmbeddingAdapter",
-    "MockDesktopAdapter",
-    "PyAutoGUIAdapter",
-]
+__all__ = ["LocalFileSystemAdapter", "MockEmbeddingAdapter", "OpenAIEmbeddingAdapter",
+           "MockDesktopAdapter", "PyAutoGUIAdapter", "RuleBasedAgentAdapter"]